Leitura de Números Com Arrays

+++ Básicos/Média entre números/media.C
@@ -7,7 +7,7 @@ int main (){
 int primeiroNumero;
 int segundoNumero;
 int terceiroNumero;
-float media;
+int media;
     
 printf("Digite o primeiro numero: \n");
     scanf("%d", &primeiroNumero);
@@ -18,7 +18,7 @@ printf("Digite o primeiro numero: \n");
     printf("Digite o terceiro numero: \n");
     scanf("%d", &terceiroNumero);
 
-    media = (primeiroNumero + segundoNumero + terceiroNumero) / 3.0;
-    printf("A media de %d, %d e %d = %.2f", primeiroNumero, segundoNumero, terceiroNumero, media);
+    media = (primeiroNumero + segundoNumero + terceiroNumero) / 3;
+    printf("A media de %d, %d e %d = %d", primeiroNumero, segundoNumero, terceiroNumero, media);
     return 0;
 }
